src/deployment/requirements.txt
# Python dependencies for ETL import script
pandas>=2.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.0.0  # Required for reading Excel files